Add files via upload
The application for creating reservations in a restaurant and viewing existing ones. The homepage of the app offers options to view the restaurant details screen. The "Reservation" button takes us to the Reservation screen where we enter the reservation details (first name, last name, time, date, number of people, etc.). The "List" button takes us to the List screen where we can see a list of existing reservations. By clicking on a name from the list, the Details screen opens, showing the details of the specific reservation.